finish part4, tests for DELETE and PUT are not required so they are commented out

+++ backend/controllers/blogs.js
@@ -2,18 +2,12 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', {
+    username: 1,
+    name: 1,
+    id: 1,
+  })
   response.json(blogs)
-})
-
-blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
-
-  if (!blog.title || !blog.url) response.status(400).end()
-  blog.likes ? blog.likes : (blog.likes = 0)
-
-  const savedBlog = await blog.save()
-  response.status(201).json(savedBlog)
 })
 
 blogsRouter.get('/:id', async (request, response) => {
@@ -21,7 +15,47 @@ blogsRouter.get('/:id', async (request, response) => {
   blog ? response.json(blog) : response.status(404).end()
 })
 
+blogsRouter.post('/', async (request, response) => {
+  const body = request.body
+  const user = request.user
+
+  if (!request.token) {
+    response.status(401).end()
+  }
+
+  if (!body.title || !body.url) {
+    response.status(400).end()
+  }
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes ? body.likes : 0,
+    user: user.id,
+  })
+
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.status(201).json(savedBlog)
+})
+
 blogsRouter.delete('/:id', async (request, response) => {
+  // const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  // if (!decodedToken) {
+  //   return response.status(401).json({ error: 'token invalid' })
+  // }
+  // const user = await User.findById(decodedToken.id)
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response
+      .status(401)
+      .json({ error: 'Blog can only be removed by its owner' })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
